# Code cleanup and documentation - Checked for style and formatting - Added XML documentation for future reference.
