+++ Assets/Scripts/GUI/PostProcessFX.cs
@@ -4,8 +4,25 @@ using UnityEngine;
 using UnityEngine.Rendering;
 using UnityEngine.Rendering.Universal;
 
+/// <summary>
+/// Utility for manipulating Post Processing FX routines for dynamic VFX.
+/// <list type="bullet">
+/// <item>
+/// <term>Focal Length</term>
+/// <description>LerpFocalLength interpolates Focal length settings
+/// to provide depth of view effects relative to phone placement.</description>
+/// </item>
+/// <item>
+/// <term>Danger Effects</term>
+/// <description>LerpDanger interpolates between configured presets for "safe"
+/// and "danger" Post Processing states.</description>
+/// </item>
+/// <item>More to Come..?</item>
+/// </list>
+/// </summary>
 public class PostProcessFX : MonoBehaviour
 {
+    // Config Params ------------------
     [Header("Depth of Field Controls")]
     [SerializeField]
     int focalLengthRest = 1;
@@ -21,12 +38,12 @@ public class PostProcessFX : MonoBehaviour
     [SerializeField]
     float chromAbbDanger = 1f;
 
+    // Component References -----------
     Volume postFX;
     DepthOfField depthOfField;
     ChromaticAberration chromAbb;
     SplitToning splitTone;
 
-    // Start is called before the first frame update
     void Start()
     {
         postFX = GetComponent<Volume>();
@@ -37,7 +54,7 @@ public class PostProcessFX : MonoBehaviour
     }
 
     /// <summary>
-    /// <para>Lerps from "Danger" VFX state *towards* "Base" VFX state</para>
+    /// <para>Linearly interpolates from "Danger" VFX state *towards* "Base" VFX state</para>
     /// <para>Higher 't' values balance more into the Base VFX state.</para>
     /// </summary>
     /// <param name="t">time value along Lerp</param>
@@ -47,6 +64,10 @@ public class PostProcessFX : MonoBehaviour
         splitTone.balance.value = Mathf.Lerp(splitToneDanger, splitToneBase, t);
 	}
 
+    /// <summary>
+    /// <para>Linearly interpolates between resting focal length and menu focal length by t.</para>
+    /// </summary>
+    /// <param name="t">time value along Lerp</param>
     public void LerpFocalLength(float t)
 	{
         depthOfField.focalLength.value = Mathf.Lerp(focalLengthRest, focalLengthMenu, t);
