+++ Assets/Scripts/PlayerSpecific/PlayerDanger.cs
@@ -3,8 +3,13 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
+/// <summary>
+/// Handles Danger Detection for player, influencing Post Processing FX
+/// and triggering death routines.
+/// </summary>
 public class PlayerDanger : MonoBehaviour
 {
+	// Config Params ------------------
 	[SerializeField]
 	float dangerDetectRange = 15f;
 	[SerializeField]
@@ -14,16 +19,16 @@ public class PlayerDanger : MonoBehaviour
 	[SerializeField]
 	LayerMask dangerMask;
 
+	// Internal Vars ------------------
 	float currentShortestDistance;
-
 	public bool isDead = false;
 
+	// Component References -----------
 	PostProcessFX ppFX;
 	Rigidbody rb;
 	CapsuleCollider capsuleCollider;
 	CharacterController controller;
 
-	// Start is called before the first frame update
 	void Start()
 	{
 		ppFX = FindObjectOfType<PostProcessFX>();
@@ -33,10 +38,12 @@ public class PlayerDanger : MonoBehaviour
 		currentShortestDistance = dangerDetectRange;
 	}
 
-	// Update is called once per frame
+	// If player is dead, return. Else, detect danger. if an enemy is within 
+	// fatal range, kill player.
 	void FixedUpdate()
 	{
-		if (isDead) return;
+		if (isDead) 
+			return;
 
 		DetectDanger();
 
@@ -44,6 +51,10 @@ public class PlayerDanger : MonoBehaviour
 			KillPlayer();
 	}
 
+	/// <summary>
+	/// <para>Triggers player death. Toggles colliders and rigidbodies, allowing the 
+	/// player to fall. Then starts the Game Over UI fade-in co-routine.</para>
+	/// </summary>
 	void KillPlayer()
 	{
 		isDead = true;
@@ -53,6 +64,13 @@ public class PlayerDanger : MonoBehaviour
 		FindObjectOfType<GameOverMenu>().StartCoroutine("FadeInGameOver");
 	}
 
+	/// <summary>
+	/// <para>Casts an overlap sphere from the Players position applying the 
+	/// Enemy Mask to skip irrelevant hits. It then checks the results searching 
+	/// for the closest enemy, updating the Current Shortest Distance member
+	/// for the new closest enemy, or increments back up to max (Detect Range).</para>
+	/// <para>Ignores dead enemies.</para>
+	/// </summary>
 	private void DetectDanger()
 	{
 		float lastShortestDistance = currentShortestDistance;
@@ -62,6 +80,8 @@ public class PlayerDanger : MonoBehaviour
 		if (hits.Length != 0)
 			foreach (var hit in hits)
 			{
+				if (hit.GetComponent<EnemyHealth>().isDead)
+					return;
 				if (Vector3.Distance(transform.position, hit.transform.position)-1 < currentShortestDistance)
 					currentShortestDistance = Vector3.Distance(transform.position, hit.transform.position)-1;
 			}
